requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
aiohttp==3.9.1
requests==2.31.0
python-dotenv==1.0.0
pandas==2.1.4
numpy==1.24.4
scikit-learn==1.3.2
streamlit==1.28.2
psutil==5.9.6
python-multipart==0.0.6